refactor: replace old web share modal by the new one
