+++ src/features/share/pages/WebShareModal.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent } from 'react'
 
-import { WebShareModalProps } from 'features/share/types'
+import { WebShareModalProps } from 'features/share/pages/WebShareModal.web'
 
 export const WebShareModal: FunctionComponent<WebShareModalProps> = () => null
