+++ src/features/share/pages/WebShareModal.web.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import styled from 'styled-components/native'
 
-import { WebShareModalProps } from 'features/share/types'
 // We are in a .web file, with a specific behavior depending on the device
 // eslint-disable-next-line no-restricted-imports
 import {
@@ -10,6 +9,7 @@ import {
   isMobileDeviceDetectOnWeb,
 } from 'libs/react-device-detect'
 import { SocialButton } from 'libs/share/SocialButton'
+import { ShareContent } from 'libs/share/types'
 import { useCopyToClipboard } from 'libs/useCopyToClipboard/useCopyToClipboard'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
@@ -28,13 +28,20 @@ import { Twitter } from 'ui/svg/icons/socialNetwork/Twitter'
 import { WhatsApp } from 'ui/svg/icons/socialNetwork/WhatsApp'
 import { getSpacingString, Spacer } from 'ui/theme'
 
+export interface WebShareModalProps {
+  visible: boolean
+  headerTitle: string
+  shareContent: ShareContent
+  dismissModal: () => void
+}
+
 export const WebShareModal = ({
   visible,
   headerTitle,
   shareContent,
   dismissModal,
 }: WebShareModalProps) => {
-  const { message, url = '' } = shareContent
+  const { subject, body, url } = shareContent
   const socialButtonProps = [
     {
       label: 'Facebook',
@@ -43,16 +50,18 @@ export const WebShareModal = ({
         url:
           'https://www.facebook.com/sharer/sharer.php?' +
           'u=' +
-          encodeURIComponent(url) +
+          encodeURIComponent(url.toString()) +
           '&quote=' +
-          encodeURIComponent(message),
+          encodeURIComponent(body),
       },
     },
     {
       label: 'Twitter',
       icon: Twitter,
       externalNav: {
-        url: `https://twitter.com/intent/tweet?text=${message}&url=${encodeURIComponent(url)}`,
+        url: `https://twitter.com/intent/tweet?text=${body}&url=${encodeURIComponent(
+          url.toString()
+        )}`,
       },
     },
     {
@@ -62,7 +71,7 @@ export const WebShareModal = ({
         url:
           (isDesktopDeviceDetectOnWeb
             ? 'https://api.whatsapp.com/send?text='
-            : 'whatsapp://send?text=') + encodeURIComponent(message + '\n' + url),
+            : 'whatsapp://send?text=') + encodeURIComponent(`${body}\n${url}`),
       },
     },
     {
@@ -70,14 +79,14 @@ export const WebShareModal = ({
       icon: Telegram,
       externalNav: {
         url: isDesktopDeviceDetectOnWeb
-          ? `https://telegram.me/share/msg?url=${encodeURIComponent(url)}&text=${message}`
-          : `tg://msg?text=${encodeURIComponent(message + '\n' + url)}`,
+          ? `https://telegram.me/share/msg?url=${encodeURIComponent(url.toString())}&text=${body}`
+          : 'tg://msg?text=' + encodeURIComponent(`${body}\n${url}`),
       },
     },
   ]
 
   const onCopyPress = useCopyToClipboard({
-    textToCopy: url,
+    textToCopy: url.toString(),
     snackBarMessage: 'Le lien a été copié dans le presse-papier\u00a0!',
     onCopy: dismissModal,
   })
@@ -107,7 +116,11 @@ export const WebShareModal = ({
           <NonSocialButtonsItem>
             <ExternalTouchableLink
               as={ButtonTertiaryBlack}
-              externalNav={{ url: 'mailto:' + '' + '?subject=' + message + '&body=' + url }}
+              externalNav={{
+                url: `mailto:?subject=${subject || body}&body=${encodeURIComponent(
+                  `${body}\n${url}`
+                )}`,
+              }}
               wording="E-mail"
               accessibilityLabel="Ouvrir le gestionnaire mail"
               icon={EmailFilled}
@@ -119,7 +132,11 @@ export const WebShareModal = ({
               <NonSocialButtonsItem>
                 <ExternalTouchableLink
                   as={ButtonTertiaryBlack}
-                  externalNav={{ url: `sms:${chooseContact}?&body=${message}: ${url}` }}
+                  externalNav={{
+                    url: `sms:${chooseContact}?&body=${body}: ${encodeURIComponent(
+                      url.toString()
+                    )}`,
+                  }}
                   wording="SMS"
                   accessibilityLabel="Ouvrir l’application de message"
                   icon={SMSFilled}
