--- src/features/share/pages/WebShareModalBest.tsx
@@ -1,5 +1,0 @@
-import { FunctionComponent } from 'react'
-
-import { WebShareModalProps } from 'features/share/pages/WebShareModalBest.web'
-
-export const WebShareModal: FunctionComponent<WebShareModalProps> = () => null
