+++ src/features/venue/components/VenueHeader/VenueHeader.tsx
@@ -6,7 +6,7 @@ import { VenueResponse } from 'api/gen'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
 import { getShareVenue } from 'features/share/helpers/getShareVenue'
-import { WebShareModal } from 'features/share/pages/WebShareModalBest'
+import { WebShareModal } from 'features/share/pages/WebShareModal'
 import { analytics } from 'libs/analytics'
 import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
 import { RoundedButton } from 'ui/components/buttons/RoundedButton'
